Product Show / Base de Datos
En la base de datos cambie todos los decimal por integer. Porq laravel no acepta decimal, solo float. Le agregue funcionalidad a la mayoria de las partes de porduct show. Modifique varias cosas a medida q iba necesitando

+++ database/migrations/2024_05_15_170040_create_product_items_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained();
             
-            $table->decimal('original_price');
-            $table->decimal('sale_price')->nullable();
+            $table->integer('original_price');
+            $table->integer('sale_price')->nullable();
             $table->foreignId('color_id')->constrained();
             $table->timestamps();
         });
